+++ satella/os/__init__.py
@@ -1,10 +1,10 @@
 from .daemon import daemonize
-from .misc import suicide, is_running_as_root
+from .misc import suicide, is_running_as_root, whereis
 from .pidlock import PIDFileLock
 from .signals import hang_until_sig
 
 __all__ = [
-    'daemonize',
+    'daemonize', 'whereis',
     'PIDFileLock', 'hang_until_sig',
     'suicide', 'is_running_as_root'
 ]