+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.16.7
 
 * added `MetricDataCollection.remove_internals`
+* added `whereis`