+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `MetricDataCollection.remove_internals`
 * added `whereis`
+* added `timeout` to `ThreadCollection.join`