+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.16.8
+
+* added extra attribute to `Future`s returned
+    by call_in_separate_thread