+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# v2.16.8
+# v2.17
 
 * added extra attribute to `Future`s returned
     by call_in_separate_thread
+* fixed a bug with `trace_function's` `tags_factory`
+
+WARNING!! Slight API change there.